Reformat script and add link to md file
